Add Request Guards
Non logged in users can no longer hard request a route. Logged in users can no longer access Login, or Register routes without logging out first. Cookies are no longer used in route functions, and instead are only accesed in the Request Guard that parses it and determines if the user is logged in, and then passes the User Id to the function for further use (querying DB).