Categories must be confirmed to be deleted.
I'm still not super happy with it, because it almost feels like too much. I might just change it so the form is just a button that says are you sure? And if they click that, it will delete it for good, so they don't have to write down the expense_id cause that's a kind of weird thing to do.